# frontend/openapi.json
{
  "components": {
    "schemas": {
      "Artifact": {
        "properties": {
          "asset_url": {
            "title": "Asset Url",
            "type": "string"
          },
          "kind": {
            "enum": [
              "audio",
              "subtitle",
              "markdown",
              "image",
              "json",
              "zip",
              "log"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "label",
          "path",
          "kind",
          "asset_url"
        ],
        "title": "Artifact",
        "type": "object"
      },
...
      "FrameCandidateIndex": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/FrameCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "chapter_contexts": {
            "items": {
              "$ref": "#/components/schemas/FrameCandidateChapterContext"
            },
            "title": "Chapter Contexts",
            "type": "array"
          }
        },
        "title": "FrameCandidateIndex",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobPublicState": {
        "properties": {
          "artifact_revision": {
            "default": "",
            "title": "Artifact Revision",
            "type": "string"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/Artifact"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "download_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Filename"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "failure_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FailureContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "progress": {
            "title": "Progress",
            "type": "integer"
          },
          "stage": {
            "$ref": "#/components/schemas/JobStage",
            "default": "queued"
          },
          "stage_elapsed_seconds": {
            "default": 0,
            "title": "Stage Elapsed Seconds",
            "type": "number"
          },
          "state_revision": {
            "default": 0,
            "minimum": 0.0,
            "title": "State Revision",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "step": {
            "title": "Step",
            "type": "string"
          },
          "step_started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Started At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "work_progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranscriptionWorkProgress"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "job_id",
          "status",
          "step",
          "progress"
        ],
        "title": "JobPublicState",
        "type": "object"
      },
      "JobStage": {
        "enum": [
          "queued",
          "analyzing_video",
          "extracting_audio",
          "transcribing",
          "awaiting_subtitle_review",
          "generating_note",
          "generating_frames",
          "preparing_review",
          "awaiting_note_review",
          "finalizing",
          "completed",
          "failed",
          "cancelling",
          "cancelled"
        ],
        "title": "JobStage",
        "type": "string"
      },
      "JobStatus": {
        "enum": [
          "pending",
          "running",
          "cancelling",
          "awaiting_subtitle_confirmation",
          "awaiting_note_review",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "QualityReport": {
        "properties": {
          "chapter_reports": {
            "items": {
              "$ref": "#/components/schemas/ChapterQualityReport"
            },
            "title": "Chapter Reports",
            "type": "array"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/QualityIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "scores": {
            "$ref": "#/components/schemas/QualityScores"
          },
          "status": {
            "enum": [
              "ready",
              "review_recommended",
              "needs_attention"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "scores"
        ],
        "title": "QualityReport",
        "type": "object"
      },
...
      "ReviewAssets": {
        "properties": {
          "frame_candidates": {
            "$ref": "#/components/schemas/FrameCandidateIndex"
          },
          "quality_report": {
            "$ref": "#/components/schemas/QualityReport"
          },
          "review_draft": {
            "$ref": "#/components/schemas/ReviewDraft"
          }
        },
        "required": [
          "frame_candidates",
          "quality_report",
          "review_draft"
        ],
        "title": "ReviewAssets",
        "type": "object"
      },
      "ReviewDraft": {
        "properties": {
          "finalized_note_sha256": {
            "default": "",
            "title": "Finalized Note Sha256",
            "type": "string"
          },
          "note_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Version Id"
          },
          "paragraphs": {
            "items": {
              "$ref": "#/components/schemas/ReviewDraftParagraph"
            },
            "title": "Paragraphs",
            "type": "array"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "source_note_sha256": {
            "default": "",
            "title": "Source Note Sha256",
            "type": "string"
          },
          "source_transcript_sha256": {
            "default": "",
            "title": "Source Transcript Sha256",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "title": "ReviewDraft",
        "type": "object"
      },
...
      "ReviewDraftParagraphUpdate": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "selected_frame_ids": {
            "items": {
              "type": "string"
            },
            "title": "Selected Frame Ids",
            "type": "array"
          },
          "status": {
            "default": "edited",
            "enum": [
              "needs_review",
              "edited",
              "approved"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "ReviewDraftParagraphUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Video Note Generator MVP",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs/{job_id}/finalize": {
      "post": {
        "operationId": "finalize_job_api_jobs__job_id__finalize_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version Id"
            }
          },
          {
            "in": "query",
            "name": "expected_state_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected State Revision"
            }
          },
          {
            "in": "query",
            "name": "expected_artifact_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Artifact Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPublicState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finalize Job",
        "tags": [
          "review"
        ]
      }
    },
    "/api/jobs/{job_id}/frame-candidates": {
      "get": {
        "operationId": "get_frame_candidates_api_jobs__job_id__frame_candidates_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameCandidateIndex"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Frame Candidates",
        "tags": [
          "review"
        ]
      }
    },
    "/api/jobs/{job_id}/frame-candidates/{candidate_id}/reject": {
      "post": {
        "operationId": "reject_job_frame_candidate_api_jobs__job_id__frame_candidates__candidate_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_state_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected State Revision"
            }
          },
          {
            "in": "query",
            "name": "expected_artifact_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Artifact Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameCandidateIndex"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Job Frame Candidate",
        "tags": [
          "review"
        ]
      }
    },
    "/api/jobs/{job_id}/frame-candidates/{candidate_id}/select": {
      "post": {
        "operationId": "select_job_frame_candidate_api_jobs__job_id__frame_candidates__candidate_id__select_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_state_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected State Revision"
            }
          },
          {
            "in": "query",
            "name": "expected_artifact_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Artifact Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameCandidateIndex"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Job Frame Candidate",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/jobs/{job_id}/quality-report": {
      "get": {
        "operationId": "get_quality_report_api_jobs__job_id__quality_report_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Quality Report",
        "tags": [
          "review"
        ]
      }
    },
    "/api/jobs/{job_id}/review-assets/prepare": {
      "post": {
        "operationId": "prepare_job_review_assets_api_jobs__job_id__review_assets_prepare_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewAssets"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prepare Job Review Assets",
        "tags": [
          "review"
        ]
      }
    },
    "/api/jobs/{job_id}/review-draft": {
      "get": {
        "operationId": "get_job_review_draft_api_jobs__job_id__review_draft_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Review Draft",
        "tags": [
          "review"
        ]
      }
    },
    "/api/jobs/{job_id}/review-draft/paragraphs/{paragraph_id}": {
      "patch": {
        "operationId": "update_job_review_draft_paragraph_api_jobs__job_id__review_draft_paragraphs__paragraph_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "paragraph_id",
            "required": true,
            "schema": {
              "title": "Paragraph Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version Id"
            }
          },
          {
            "in": "query",
            "name": "expected_state_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected State Revision"
            }
          },
          {
            "in": "query",
            "name": "expected_artifact_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Artifact Revision"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDraftParagraphUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Job Review Draft Paragraph",
        "tags": [
          "review"
        ]
      }
    },
...
  }
}